Structure de communication
La communication API, le builder et le traitement JSON ont été séparés
Le builder a changé
De vieux fichiers inutilisés ont été supprimés
