--- src/fr/groupe4/clientprojet/communication/CommunicationType.java
@@ -1,8 +1,0 @@
-package fr.groupe4.clientprojet.communication;
-
-public enum CommunicationType {
-    LOGIN,
-    CHECK_CONNECTION,
-    UPDATE_CONNECTION,
-    LIST_PROJECTS;
-}
