--- src/fr/groupe4/clientprojet/communication/TaskList.java
@@ -1,12 +1,0 @@
-package fr.groupe4.clientprojet.communication;
-
-import java.util.ArrayList;
-
-/**
- * Liste des tâches
- *
- * @author Romain
- * @see Task
- */
-public class TaskList extends ArrayList<Task> {
-}
