+++ src/fr/groupe4/clientprojet/communication/enums/CommunicationType.java
@@ -1,0 +1,11 @@
+package fr.groupe4.clientprojet.communication.enums;
+
+/**
+ * Types de communication
+ */
+public enum CommunicationType {
+    LOGIN,
+    CHECK_CONNECTION,
+    UPDATE_CONNECTION,
+    LIST_PROJECTS
+}
