+++ src/fr/groupe4/clientprojet/display/dialog/connectiondialog/controller/EventConnectionDialog.java
@@ -45,10 +45,10 @@ public class EventConnectionDialog extends WindowAdapter implements ActionListen
     public void actionPerformed(ActionEvent e) {
         switch (ConnectionChoice.getEnum(e.getActionCommand())) {
             case OK:
-                Communication comm = new Communication.CommunicationBuilder()
-                                                      .startNow()
-                                                      .connect(source.getUsername(), source.getPassword())
-                                                      .build();
+                Communication comm = Communication.builder()
+                                                  .startNow()
+                                                  .connect(source.getUsername(), source.getPassword())
+                                                  .build();
                 new LoadDialog(source, comm);
 
                 if (Communication.isConnected()) {
